:art: https://github.com/siyuan-note/siyuan/issues/12009
